[Core] [Shop] [Admin] Move listeners to proper place
