+++ src/Sylius/Bundle/AdminBundle/Resources/config/services.xml
@@ -15,18 +15,13 @@
     <imports>
         <import resource="services/controller.xml" />
         <import resource="services/menu.xml" />
+        <import resource="services/listener.xml" />
     </imports>
 
     <services>
         <service id="sylius.context.locale.admin_based" class="Sylius\Bundle\AdminBundle\Context\AdminBasedLocaleContext">
             <argument type="service" id="security.token_storage" />
             <tag name="sylius.context.locale" priority="128" />
-        </service>
-
-        <service id="sylius.event_subscriber.resource_delete" class="Sylius\Bundle\AdminBundle\EventListener\ResourceDeleteSubscriber">
-            <argument type="service" id="router" />
-            <argument type="service" id="session" />
-            <tag name="kernel.event_subscriber" event="kernel.exception" />
         </service>
     </services>
 </container>
