+++ src/Sylius/Bundle/ShopBundle/spec/EventListener/OrderCompleteListenerSpec.php
@@ -9,11 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\CoreBundle\EventListener;
+namespace spec\Sylius\Bundle\ShopBundle\EventListener;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Bundle\CoreBundle\EmailManager\OrderEmailManagerInterface;
-use Sylius\Bundle\CoreBundle\EventListener\OrderCompleteListener;
+use Sylius\Bundle\ShopBundle\EventListener\OrderCompleteListener;
 use Sylius\Component\Core\Model\OrderInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
